+++ app/Config/Routes.php
@@ -23,6 +23,18 @@ $routes->get('mahasiswa-lulus-do', 'MahasiswaLulusDo::index');
 
 /*
  * --------------------------------------------------------------------
+ * Neo Feeder CRUD Routes (ENH-015)
+ * --------------------------------------------------------------------
+ */
+$routes->get('mahasiswa/edit/(:any)', 'Mahasiswa::edit');
+$routes->post('mahasiswa/edit/(:any)', 'Mahasiswa::editPost');
+$routes->post('mahasiswa/delete/(:any)', 'Mahasiswa::delete');
+$routes->get('aktivitas-kuliah/edit/(:any)/(:any)', 'AktivitasKuliah::edit');
+$routes->post('aktivitas-kuliah/edit/(:any)/(:any)', 'AktivitasKuliah::editPost');
+$routes->post('aktivitas-kuliah/delete/(:any)/(:any)', 'AktivitasKuliah::delete');
+
+/*
+ * --------------------------------------------------------------------
  * PISN Graduation Wizard Routes (ENH-013)
  * --------------------------------------------------------------------
  */